1 remove .idea from repository 2 hibernate add prop validate tables 3 fix but float double

+++ src/sql/create_tables.sql
@@ -4,8 +4,8 @@ USE auction;
 
 DROP TABLE IF EXISTS users;
 CREATE TABLE IF NOT EXISTS users (
-  id      INT PRIMARY KEY AUTO_INCREMENT,
-  email   VARCHAR(128) NOT NULL,
+  id       INT PRIMARY KEY AUTO_INCREMENT,
+  email    VARCHAR(128) NOT NULL,
   password VARCHAR(128) NOT NULL
 );
 
@@ -14,7 +14,7 @@ DROP TABLE IF EXISTS items;
 CREATE TABLE IF NOT EXISTS items (
   id      INT PRIMARY KEY AUTO_INCREMENT,
   name    VARCHAR(128) NOT NULL,
-  price   FLOAT        NOT NULL,
+  price   DOUBLE       NOT NULL,
   user_id INT          NOT NULL,
   FOREIGN KEY (user_id) REFERENCES users (id)
 );
